refactor: ajustando camada de repository & datasource conforme documentação: - removendo implementação da camada de repository e alterando interface de implementação da datasource - ajustando pacote e nome da classe para que seja agnostico a tecnologia (Datasource, ORM e e pacote)
